grid dataset

+++ torchcv/models/retinanet/box_coder.py
@@ -5,6 +5,8 @@ import torch
 from torchcv.utils import meshgrid
 from torchcv.utils.box import box_iou, box_nms, change_box_order
 
+import pdb
+
 
 class RetinaBoxCoder:
     def __init__(self):
@@ -51,7 +53,7 @@ class RetinaBoxCoder:
             grid_size = input_size / fm_size
             fm_w, fm_h = int(fm_size[0]), int(fm_size[1])
             xy = meshgrid(fm_w,fm_h) + 0.5  # [fm_h*fm_w, 2]
-            xy = (xy*grid_size).view(fm_h,fm_w,1,2).expand(fm_h,fm_w,9,2)
+            xy = (xy*grid_size.float()).view(fm_h,fm_w,1,2).expand(fm_h,fm_w,9,2)
             wh = anchor_wh[i].view(1,1,9,2).expand(fm_h,fm_w,9,2)
             box = torch.cat([xy-wh/2.,xy+wh/2.], 3)  # [x,y,x,y]
             boxes.append(box.view(-1,4))
